requirements.txt
praw
google-generativeai
psycopg2-binary
python-dotenv
